remove "network_ai" and "network" controllers
and instead have a field "is_networked" in [side].

This might break existent wml that uses side controller= to check
whether a side is networked, the plan is to add a lua function
wesnoth.is_networked(side) if someone needs it. These changes in wml syntax
are not side effects, they are inteded. The reason is, the wml dev shouldn't
worry about whether a side is networked or not. Especially we remove the OOS
source in [store_side]. It also makes makes it easier to play Sp scenarios
(that do  checks like controller == "human" controllers) in mp, and it
disallows stuff like [modify_side] controller="network" which always causes
bugs.

This also simplifies that servercode that previously changed the
"controller" value a lot and had to do check for every value of controller=
now it's more easy becasue he just needs to set the "is_networked" flag.

Conflicts:
	src/saved_game.cpp
	src/server/game.cpp