fix mp scenario transitions bugged in scenarios that change controllers
fixes http://gna.org/bugs/?23646 and also fixes scenario transitions to
scenarios with_allow_new_game =yes (previously users were kicked to the
Lobby in that case)

As commented in the previous code , it had some problems: It used the
current side controllers information to maniputlte the starting point
for players who just started this scenario (by advancing from the
previous scenario). This is incorrect if that controllers information
has changed sonce the start of the scenario.

The new beahviour is that players from the previous scenario just get
the observers starting point and get controller updates later
.