refactor play_controller variables and game_data::phase
Instead of having variables init_side_done_, linger_ and the
END_TURN_STATE enum, we not have more member in game data::PHASE
and a boolean ent_turn_required_ variable.

The intention is to make the code simpler to prepare a refactor
to move the linger code into the play_scenario function so that
the replay will be able to execute victory events.

Its still a bit arbitrary which variables are part of game_data
or game_state imo.