mark campaign as finished for losses with proceed_to_next_level
unless explicitly specified proceed_to_next_levelmatches is_victory
and we already checked for proceed_to_next_level before. I think it
makes a lot of sense to only depend on proceed_to_next_level and
not on is_victory here. This gives users the theoretical option to
end the campaign with a defeat, but more importantly it allowseasier code.