add a synced_context::block_undo() function
Using this function lets te synced context know that the current
action cannot be undone, so that it will in particular send the
actions too all players as soon as possible. The intention is
less delay in networked games.

Currently there is some redundance between is_simultanious_
and is_undo_blocked_, maybe one of these will be removed
later.